store.js moved to core, main saga created, watchFetchMoviesData rename to movieSaga
