+++ src/core/saga.js
@@ -1,0 +1,8 @@
+import { all } from "redux-saga/effects";
+import { movieSaga } from "../features/movieSaga";
+
+export default function* saga() {
+  yield all([
+    movieSaga(),
+  ]);
+}
